Fixed pedantic warning about unreliable floating point comparison in case a floating point number is passed into the LIBXSMM_NBITS macro (perform explicit conversion to integer).

+++ include/libxsmm_macros.h
@@ -155,7 +155,7 @@
 #define LIBXSMM_NBITS16(N) (0 != ((N) & 0xFF00) ? (8 + LIBXSMM_NBITS08((N) >> 8)) : LIBXSMM_NBITS08(N))
 #define LIBXSMM_NBITS32(N) (0 != ((N) & 0xFFFF0000) ? (16 + LIBXSMM_NBITS16((N) >> 16)) : LIBXSMM_NBITS16(N))
 #define LIBXSMM_NBITS64(N) (0 != ((N) & 0xFFFFFFFF00000000) ? (32 + LIBXSMM_NBITS32((N) >> 32)) : LIBXSMM_NBITS32(N))
-#define LIBXSMM_NBITS(N) (0 != (N) ? (LIBXSMM_NBITS64((unsigned long long)(N)) + 1) : 1)
+#define LIBXSMM_NBITS(N) (0 != ((unsigned long long)(N)) ? (LIBXSMM_NBITS64((unsigned long long)(N)) + 1) : 1)
 
 #define LIBXSMM_DEFAULT(DEFAULT, VALUE) (0 < (VALUE) ? (VALUE) : (DEFAULT))
 #define LIBXSMM_ABS(A) (0 <= (A) ? (A) : -(A))
